added authentication with JWT state change endpoints are protected by role-based authorization

+++ server/src/main/java/com/john/price/PetAdoption/Repositories/ApplicationUserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.john.price.PetAdoption.Models.ApplicationUser;
 
-public interface ApplicationUserRepository extends JpaRepository<ApplicationUser, Long>{
+public interface ApplicationUserRepository extends JpaRepository<ApplicationUser, Integer>{
 	public ApplicationUser findByUsername(String username);
 }
